pyproject.toml
[tool.poetry]
name = "webp2png"
version = "0.1.0"
description = ""
authors = ["zhaifengtwo <zhaifengtwo@gmail.com>"]
packages = [{include="webp2png"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
